Create and restructure popup component
