+++ src/app/grant/grant-details/grant-details-popup/grant-detail-popup.component.html
@@ -1,0 +1,21 @@
+<h1 mat-dialog-title>Delete Grant Profile</h1>
+<div mat-dialog-content>
+  <p>Are you sure you would like to delete this Grant?</p>
+  <button
+    style="margin-right: 20px"
+    mat-button
+    class="grant-btn"
+    type="submit"
+    (click)="closeDialog()"
+  >
+    Cancel
+  </button>
+  <button
+    mat-button
+    class="grant-btn delete-btn"
+    type="submit"
+    (click)="deleteGrant(grandId)"
+  >
+    Confirm
+  </button>
+</div>
