+++ src/app/grant/grant-details/grant-details.component.ts
@@ -8,6 +8,7 @@ import { ActivatedRoute, Router } from "@angular/router";
 import { ToastrService } from "ngx-toastr";
 import { Grant } from "../shared/grant.model";
 import { GrantService } from "../shared/grant.service";
+import { GrantDetailsPopupConponent } from "./grant-details-popup/grant-detail-popup.component";
 
 export interface DialogData {
   grand_id: number;
@@ -55,49 +56,49 @@ export class GrantDetailsComponent implements OnInit {
   }
 }
 
-@Component({
-  selector: "grant-details-popup",
-  templateUrl: "grant-details-popup.html",
-  styleUrls: ["./grant-details.component.scss"],
-})
-export class GrantDetailsPopupConponent {
-  grandId: any;
+// @Component({
+//   selector: "grant-details-popup",
+//   templateUrl: "grant-details-popup.html",
+//   styleUrls: ["./grant-details.component.scss"],
+// })
+// export class GrantDetailsPopupConponent {
+//   grandId: any;
 
-  constructor(
-    public dialog: MatDialog,
-    private toastr: ToastrService,
-    private router: Router,
-    private grantService: GrantService,
+//   constructor(
+//     public dialog: MatDialog,
+//     private toastr: ToastrService,
+//     private router: Router,
+//     private grantService: GrantService,
 
-    public dialogRef: MatDialogRef<GrantDetailsPopupConponent>,
-    @Inject(MAT_DIALOG_DATA) public data: DialogData
-  ) {}
+//     public dialogRef: MatDialogRef<GrantDetailsPopupConponent>,
+//     @Inject(MAT_DIALOG_DATA) public data: DialogData
+//   ) {}
 
-  closeDialog() {
-    this.dialog.closeAll();
-  }
+//   closeDialog() {
+//     this.dialog.closeAll();
+//   }
 
-  ngOnInit() {}
+//   ngOnInit() {}
 
-  deleteGrant(grantId: number) {
-    console.log("DeleteID " + this.data.grand_id);
-    grantId = this.data.grand_id;
-    console.log("Delete");
-    this.grantService.deleteGrant(grantId).subscribe(
-      (res) => {
-        console.log(res);
-        this.toastr.success("Delete Grant!", res.message, {
-          timeOut: 3000,
-        });
-        this.closeDialog();
-        this.router.navigate(["/grants"]);
-      },
-      (err) => {
-        console.log(err);
-        this.toastr.error(" Error Deleting Grant", "Server Error!", {
-          timeOut: 3000,
-        });
-      }
-    );
-  }
-}
+//   deleteGrant(grantId: number) {
+//     console.log("DeleteID " + this.data.grand_id);
+//     grantId = this.data.grand_id;
+//     console.log("Delete");
+//     this.grantService.deleteGrant(grantId).subscribe(
+//       (res) => {
+//         console.log(res);
+//         this.toastr.success("Delete Grant!", res.message, {
+//           timeOut: 3000,
+//         });
+//         this.closeDialog();
+//         this.router.navigate(["/grants"]);
+//       },
+//       (err) => {
+//         console.log(err);
+//         this.toastr.error(" Error Deleting Grant", "Server Error!", {
+//           timeOut: 3000,
+//         });
+//       }
+//     );
+//   }
+// }
